NCL-4672 Use internal loggerin analyzer/manipulator modules. Refactor code.
