+++ common/src/main/java/org/jboss/gm/common/logging/FilteringCustomLogger.java
@@ -1,4 +1,4 @@
-package org.jboss.gm.analyzer.alignment.util;
+package org.jboss.gm.common.logging;
 
 import java.util.Arrays;
 import java.util.List;
@@ -6,6 +6,8 @@ import java.util.List;
 import org.gradle.internal.logging.events.LogEvent;
 import org.gradle.internal.logging.events.OutputEvent;
 import org.gradle.internal.logging.events.OutputEventListener;
+import org.gradle.internal.logging.slf4j.OutputEventListenerBackedLoggerContext;
+import org.slf4j.LoggerFactory;
 
 public class FilteringCustomLogger implements OutputEventListener {
 
@@ -21,12 +23,18 @@ public class FilteringCustomLogger implements OutputEventListener {
 
     private final List<String> ownCategories = Arrays.asList(
             "org.jboss.gm.analyzer.alignment.AlignmentTask_Decorated",
-            "org.jboss.gm.analyzer.alignment.util.FilteringCustomLogger");
+            "org.jboss.gm.common.logging.FilteringCustomLogger");
 
     @SuppressWarnings("FieldCanBeLocal")
     private final String defaultCategory = "org.gradle.api.Task";
 
-    public FilteringCustomLogger(OutputEventListener outputEventListener) {
+    public static void enableFilter() {
+        OutputEventListenerBackedLoggerContext context = (OutputEventListenerBackedLoggerContext) LoggerFactory
+                .getILoggerFactory();
+        context.setOutputEventListener(new FilteringCustomLogger(context.getOutputEventListener()));
+    }
+
+    private FilteringCustomLogger(OutputEventListener outputEventListener) {
         this.delegate = outputEventListener;
     }
 
@@ -38,7 +46,8 @@ public class FilteringCustomLogger implements OutputEventListener {
 
             if (!ownCategories.contains(logEvent.getCategory()) &&
                     !logEvent.getCategory().equals(defaultCategory)) {
-                // Can't use logger to output the warning as causes stackoverflow.
+                // Can't use logger to output the warning as causes stack overflow.
+                // TODO: Remove?
                 System.err.println("Unknown event using category " + logEvent.getCategory() + " : ");
                 delegate.onOutput(event);
                 System.err.println(".... completed unknown event.");
