+++ common/src/main/java/org/jboss/gm/common/logging/GMLogger.java
@@ -1,4 +1,4 @@
-package org.jboss.gm.common;
+package org.jboss.gm.common.logging;
 
 import java.util.Arrays;
 
@@ -7,6 +7,7 @@ import org.commonjava.maven.ext.common.ManipulationUncheckedException;
 import org.gradle.api.logging.LogLevel;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
+import org.jboss.gm.common.Configuration;
 import org.slf4j.Marker;
 
 @SuppressWarnings("unused")
@@ -30,8 +31,8 @@ public class GMLogger implements Logger {
 
     private Configuration configuration;
 
-    public static GMLogger getLogger(String name) {
-        return new GMLogger(name);
+    public static Logger getLogger(Class c) {
+        return new GMLogger(c.getName());
     }
 
     private GMLogger(String name) {
