+++ manipulation/build.gradle.kts
@@ -32,7 +32,6 @@ dependencies {
     compile("commons-lang:commons-lang:${project.extra.get("commonsVersion")}")
     compile("org.commonjava.maven.ext:pom-manipulation-common:${project.extra.get("pmeVersion")}")
     compile("org.commonjava.maven.atlas:atlas-identities:${project.extra.get("atlasVersion")}")
-    compile("org.slf4j:slf4j-api:${project.extra.get("slf4jVersion")}")
 
     testCompile("junit:junit:${project.extra.get("junitVersion")}")
     testCompile("org.assertj:assertj-core:${project.extra.get("assertjVersion")}")
